feat(proto_verifier_test): add cc_deps and cc_indexer parameters
cc_indexer allows passing a different indexer binary, which we'll use internally for testing proto static reflection
cc_deps allows additional c++ deps to the file being tested, which we'll also be using for testing proto static reflection

Note that there may be a better way to make the changes to cc_indexer_test.bzl. It seems when allow_files is set to a list of allowed extensions, the providers list is ignored and cc_library targets are not allowed to be passed. Setting allow_files = True works around this.