chore(cxx_indexer): remove experimental_guess_proto_semantics
Our version of protobuf now contains the semantic field in
GeneratedCodeInfo, so experimental_guess_proto_semantics is
now redundant.

Our version of protobuf does *not* fill out all semantics.
The ones we expect to come in a future import are commented
out in the test. IIRC it was determined that arena maintenance
functions should not have semantics set, so these have been
removed entirely.

In service of #5705