fix(cxx_indexer): always ref a decl regardless of alternate targets
At the intersection of indirection, write refs and protobuf,
we have the `mutable_` methods (without loss of generality).
The C++ indexer can look underneath assignments through simple
indirection (`*mutable_foo() = bar`) and can also identify
the field (foo) being set. Before this PR, the indexer would not
emit a ref to the `mutable_foo` function if it found a "better"
reference target because of the indirection. Now, it will
check to see if the alternate target is different from the
ordinary referent; if so, it will emit an ordinary reference to
the ordinary referent.

This PR also adds tests demonstrating the difference in
behavior between `mutable_` as alias versus `mutable_` as
a simple write.