fix(verifier): request larger stack size for new solver
In the future we can look at doing a smarter job lowering to datalog
(e.g. not adding every goal into the same giant Horn clause, which was
causing souffle to bump up against the stack). For now, just asking for
a larger stack allows all the Go tests to pass; tested this by running
them with the flag fixed to true.