feat(verifier): warn on and allow specifying default file corpora
This PR adds a flag to specify default file vname corpus values
for anchor goals. These are chiefly useful when providing files
as additional input to the verifier with --file_vnames on (the
default). It's meant to help migrate indexers and test harnesses
that set file corpora in anchors but do not also provide them
for the corresponding file nodes (which happens surprisingly often),
causing tests to fail when anchor corpora are checked (as happens
in the fast solver, but has been "relaxed" in the old solver for
a very long time).