feat(verifier): avoid the arity ceiling in outputs (#5939)
* feat(verifier): avoid the arity ceiling in outputs

The new solver/Souffle has a hard limit on the arity of relations.
This is tunable but requires a change to the library. Before this PR,
that limit (20 as I write this) was easy to reach through inspections,
especially when `--annotated_graphviz` is on (which adds an inspection
to *every* evar). Luckily, this limit does not apply to records.
This PR changes lowering to use a unary output relation when
inspections are on. The single element is a record containing the
values of the inspected evars.