feat(verifier): print inspections with fast solver
This PR also gets rid of the compatibility API for
VerifyAllGoals that ignores the string_view (that the
fast solver fills with the pretty-printed value for the
inspection).

A future PR will change `saved_assignments_` and will
update the various graphviz printers to work with the fast
solver (again by using the pretty-printed values instead
of ast pointers, which we don't have anymore).