fix(verifier): give more useful diagnostics during recovery (#5971)
* fix(verifier): give more useful diagnostics during recovery

We might want to consider continuing recovery after lowering
fails. In practice this happens when a program with inspections
is partially lowered (because the full program can't be solved)
and some inspected evars never appear, so are never given types.
Note that we'll also need to fix the output record/expression as
those datalog variables will be unconstrained.

This PR does a simpler thing and cleans up errors during
recovery (to clarify that the 'no type found for X' messages are
expected); it also selects a more relevant error goal.