feat(cxx_extractor): better error when opening KzipWriter
Add the path to the kzip file to the error message when it fails to
open.

When using `runextractor compdb` from `compile_commands.json`
generated with `refresh_compile_commands` under Bazel 6.5.x with
default system GCC-based toolchain, extractions seem to be reporting
en-masse that `ALREADY_EXISTS: File already exists`, but it is
difficult to know what file is being written to.

This makes it easier to triage why this is happening.