Connect cross-references on object fields through a type wrapper.
This change enables cross-references on object literals typed using type
wrappers (such as Readonly, Immutable, etc.), specified as a global list.

It further allows these wrappers to perform a transformation on the Type, which
is future-proofing for cases where we may need to walk through a class
declaration to find a type with the appropriate properties.