chore(go): update go from 1.21 to 1.23 in extractor image (#6130)
Go itself requires at least 1.22.6 for bootstrapping as of https://go.dev/cl/606156. Thus it is no longer possible to use the extractor image on the Go repo.

Update to 1.23.0, as that is the latest release.

As an aside, it is unclear to me why there is a mismatch in the base image (1.20) and the installed version (1.21). gcr.io/cloud-builders/go doesn't seem to have a 1.23 tag yet.