feat: make a fully hermetic, ephemeral, and reproducible build
Makes the necessary changes to allow an ephemeral `nix` instance
to be used for providing all the build prerequisites.

The result is a hermetic, ephemeral and reproducible ("her") build.

As a result, provided that you have `bazel` installed, it is
possible to compile the entire kythe distribution with a
one-liner in shell.  It still takes a preposterous time to complete,
unless you have a 96-core machine like I do. In which case it's about
10 minutes.

If you are curious why this works, see:
https://hdlfactory.com/post/2024/04/20/nix-bazel-%EF%B8%8F/