KZipReaderTest: fix EmptyCatch error
As we upgraded google/error-prone, we included some of the recent
changes upstream into the Kythe code base. One of which was
https://github.com/google/error-prone/commit/1ff295936ad30dd83788b5f21ea3f265c3a31f66
which forces tests to use assertThrows instead of
`catch (FooExecption expected) {}` to validate exceptions.

Update KZipReaderTest to use this new pattern so that the test can
actually run. Note that testOpenAndReadSimpleKZip,
testOpenFailsForGarbageUnit and testKZipWithEmptyFileDoesNotCrash
are still failing, but it's likely because of different reasons.